Moved query features to `Query` namespace
